+++ packages/ui-core/src/containers/Param/ParamEnum/ParamEnum.tsx
@@ -1,6 +1,6 @@
 import { useRef } from 'react'
 import { NodeParamEnum } from '@hedron/engine'
-import { useEngineStore } from '@hooks/useStores'
+import { useEngineStore } from '@hooks/storeHooks'
 import { useOnNodeValueChange } from '@hooks/useOnNodeValueChange'
 import { useSubscribeToNodeValue } from '@hooks/useSubscribeToNodeValue'
 import { EnumDropdown, EnumDropdownHandle } from '@components/EnumDropdown/EnumDropdown'
