+++ packages/ui-core/src/containers/Param/ParamString/ParamString.tsx
@@ -1,7 +1,7 @@
 import { useRef } from 'react'
 import { useInterval } from 'usehooks-ts'
 import { useOnNodeValueChange } from '@hooks/useOnNodeValueChange'
-import { useEngineStoreWithContext } from '@hooks/useStores'
+import { useEngineStoreWithContext } from '@hooks/storeHooks'
 import { TextInput, TextInputHandle } from '@components/TextInput/TextInput'
 
 interface ParamNumberProps {
