+++ packages/ui-core/src/hooks/useSelectedNode.ts
@@ -1,4 +1,4 @@
-import { useAppStore } from '@hooks/useStores'
+import { useAppStore } from '@hooks/storeHooks'
 
 export const useSelectedNode = (sketchId: string | null) => {
   return useAppStore((state) => state.selectedNodes[sketchId ?? 'aux'])
