+++ packages/ui-core/src/hooks/useUpdateNodeValue.ts
@@ -1,4 +1,4 @@
-import { useEngineStore } from '@hooks/useStores'
+import { useEngineStore } from '@hooks/storeHooks'
 
 export const useUpdateNodeValue = () => {
   const updateNodeValue = useEngineStore((state) => state.updateNodeValue)
