+++ packages/ui-core/src/index.ts
@@ -21,6 +21,6 @@ export * from './components/NumberInput/NumberInput'
 export * from './components/Collapsible/Collapsible'
 export * from './containers/Param/Param'
 export { ParamNumberOptions } from './containers/Param/ParamNumber/ParamNumber'
-export * from './hooks/useStores'
+export * from './hooks/storeHooks'
 export * from './hooks/useOnSelectNode'
 export * from './hooks/useSelectedNode'
